+++ node/middlewares/updateLead.ts
@@ -1,26 +1,27 @@
 export async function updateLead(
-  { clients: { masterdata }, state, vtex: { logger } }: StatusChangeContext,
+  {
+    clients: { userAffiliation },
+    state,
+    vtex: { logger },
+  }: StatusChangeContext,
   next: () => Promise<unknown>
 ) {
-  const { client, affiliate } = state
+  const { userEmail: email, userProfileId: leadDocumentId, affiliate } = state
 
-  const clientNewLeadInfo = {
-    ...client,
+  const userNewLeadInfo = {
+    id: leadDocumentId,
+    email,
     affiliateId: affiliate,
     affiliateStartDate: new Date().toISOString(),
   }
 
   try {
-    await masterdata.updatePartialDocument({
-      dataEntity: 'CL',
-      id: client.id,
-      fields: clientNewLeadInfo,
-    })
+    await userAffiliation.saveOrUpdate(userNewLeadInfo)
   } catch (err) {
     logger.error({
       metric: 'update-lead',
       message: 'Error updating the lead',
-      info: clientNewLeadInfo,
+      info: userNewLeadInfo,
       error: err.message,
     })
     throw new Error('Error updating the lead')